# [Android] Adjust buffer size/duration for TrueHD IEC