## Description
[Android] Adjust buffer size/duration for TrueHD IEC

## Motivation and context
A few days ago I've been investigating the problems of TrueHD dropouts (in Shield), which still have a small percentage of users. Finally, after many tests, I have discovered interesting things:

The initial idea was to increase the buffer because it seems intuitive but it has been a failure (in master branch). The current buffer used is 160 ms, and any higher value performs worse. But this has given me clues of the way forward...

Tried all possible lower values:

160ms
128ms
96ms
64ms

96 and 64 are the ones that work best and at this point I didn't know which one to choose. But then, thinking about the possible backport to Matrix I have discovered something unexpected: Matrix already uses a value of 64 ms. I was aware that the Matrix code was different but at the time I did tests and I don't know why I came to the conclusion that for TrueHD the parameters were the same. I never got around to looking at this again.

From people's feedback it is known that changing some audio settings in Shield can reduce, even completely eliminate dropouts. Specific:

Audio video sync = 0 ms (default is 50 ms)

I have been testing Matrix with the default Shield settings and it works much worse than master in those same conditions BUT with the optimized Shield settings Matrix works better than master.

Bottom line: Matrix is ​​more sensitive to setting changes in Shield.

Then I have repeated the test in master branch but with 
Audio video sync = 50 ms (default)

160 ms <--- current master
128 ms
96 ms <---- best
64 ms <---- Matrix

The conclusion is clear 96 ms is the best buffer value (by far) when using the default Shield settings.

The idea is to change this in master only and wait for more feedback. In my setup it is clear that it works better but in other setups it may be different and therefore I consider it quite risky to make changes in Matrix at this point.

If with Alpha2 the consensus is reached that TrueHD works better than Matrix then the backport could be considered.

## How has this been tested?
Runtime tested Shield Pro 2019 - FW 9.0.2 (Android 11) / Denon AVR-X1600H

**Matrix 19.4  - 64 ms**
```
DEBUG <general>: Minimum size we need for stream: 98432
DEBUG <general>: Created Audiotrackbuffer with playing time of 64.083333 ms min buffer size: 196864 bytes
INFO <general>: CAESinkAUDIOTRACK::Initializing with: m_sampleRate: 192000 format: AE_FMT_S16LE (AE) method: IEC (PT) stream-type: STREAM_TYPE_TRUEHD min_buffer_size: 196864 m_frames: 6152 m_frameSize: 16 channels: 8
```

**current master - 160 ms (0.15)**
```
DEBUG <general>: Minimum size we need for stream: 98432
DEBUG <general>: Created Audiotrackbuffer with playing time of 160.208333 ms min buffer size: 492160 bytes
INFO <general>: CAESinkAUDIOTRACK::Initializing with: m_sampleRate: 192000 format: AE_FMT_S16LE (AE) method: IEC (PT) stream-type: STREAM_TYPE_TRUEHD min_buffer_size: 492160 m_frames: 6152 m_frameSize: 16 channels: 8
```

**PR - 96 ms (0.08)**
```
DEBUG <general>: Minimum size we need for stream: 98432
DEBUG <general>: Audiotrack buffer params are: period time = 32.042 ms, period size = 98432 bytes, num periods = 3
DEBUG <general>: Created Audiotrackbuffer with playing time of 96.125000 ms min buffer size: 295296 bytes
INFO <general>: CAESinkAUDIOTRACK::Initializing with: m_sampleRate: 192000 format: AE_FMT_S16LE (AE) method: IEC (PT) stream-type: STREAM_TYPE_TRUEHD min_buffer_size: 295296 m_frames: 6152 m_frameSize: 16 channels: 8
```

## What is the effect on users?
Reduce/eliminate completely TrueHD dropouts. Even without using optimal Shield settings can be achieved TrueHD without dropouts.


## Types of change
<!--- What type of change does your code introduce? Put an `x` in all the boxes that apply like this: [X] -->
- [X] **Bug fix** (non-breaking change which fixes an issue)
- [ ] **Clean up** (non-breaking change which removes non-working, unmaintained functionality)
- [X] **Improvement** (non-breaking change which improves existing functionality)
- [ ] **New feature** (non-breaking change which adds functionality)
- [ ] **Breaking change** (fix or feature that will cause existing functionality to change)
- [ ] **Cosmetic change** (non-breaking change that doesn't touch code)
- [ ] **None of the above** (please explain below)

## Checklist:
<!--- Go over all the following points, and put an `X` in all the boxes that apply like this: [X] -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [X] My code follows the **[Code Guidelines](https://github.com/xbmc/xbmc/blob/master/docs/CODE_GUIDELINES.md)** of this project 
- [ ] My change requires a change to the documentation, either Doxygen or wiki
- [ ] I have updated the documentation accordingly
- [X] I have read the **[Contributing](https://github.com/xbmc/xbmc/blob/master/docs/CONTRIBUTING.md)** document
- [ ] I have added tests to cover my change
- [ ] All new and existing tests passed