@fritsch 2022-05-21
@thexai Are you sure the root-cause is understood?

The old code basically distinguished between buffer (can be very large) and periods. What your code actually changes with the reduced buffer is the "ratio" towards the periods, so the periods get smaller.

Does it also work, if you plain double the amount of periods by e.g. reducing their size in half?

Old: [P1 P2 P3 P4 P5 P6 P7 P8] 180 ms
New: [P1' P1'' P2' P2'' P3' P3'' ... P8' P8''] 180 ms

with Pn' being Pn/2?

The periods is the crucial thing as that relates directly into the "amount" the sink needs to  be filled with data -> CPU is used to come back with data, while buffer basically only means to cover underruns properly. In ideal world the pseudo periods would also match the underlaying Android ALSA periods, same as buffer.

Though  giving the shear amount of data of an 8 channel IEC format ... It's perfectly fine to reduce the buffer time, as the size is quite fat :-)

So - both would be fine with me - given that we really understand what the issue is.