@fritsch 2022-05-21
In code (my wall of text):
Instead of:
```
c++
```

Just use:
```
c +=2;
```

in case of TrueHD and DTS-HD, which has exactly the same issue in IEC format.