@thexai 2022-05-21
Sorry, I don't quite understand your comments.

In Matrix:
32ms + 32ms = ~64ms     (two periods)
in bytes:
98432 + 98432  = 196864 bytes

PR :
32 + 32 + 32 = ~96 ms   (3 periods)
98432 + 98432 + 98432 = 295296 bytes

current master:
32 + 32 + 32 + 32 + 32 = ~160 ms   (5 periods)
98432 * 5 =  492160 bytes

basic "period" time or bytes never has changed, is 32ms and 98432 bytes


`c +=2;`  ??

current master is `c++` and PR does not change that:
https://github.com/xbmc/xbmc/blob/126c8500e157c3d81da1d56a11e18fd1280c99a1/xbmc/cores/AudioEngine/Sinks/AESinkAUDIOTRACK.cpp#L513


> Though giving the shear amount of data of an 8 channel IEC format ... It's perfectly fine to reduce the buffer time, as the size is quite fat :-)

We both agree on that  😀