@fritsch 2022-05-21
If all periods stay the same then we have another issue below, namely: out get delay method is not accurate enough and videoplayer auto corrects that.

As Android API for really accurate sync is still a pain until today, please go ahead. I would do it for DTS hd also if possible.