@thexai 2022-05-21
> I would do it for DTS hd also if possible.

Honestly with DTS-HD there are absolutely no complaints/issues and I haven't noticed anything either. It just works fine as it is now. In this PR I prefer not to touch anything else.