@DecalicatanDecalicatan 2022-07-10
Hi, 

With the Nexus alpha 2 that incluses this commit,  using a Shield pro 2019 and an avr (Marantz NR 1711), should the Audio video sync setting of the Shield left at 50 ms (default) or the 0ms value is still the "best"  to prevent dropouts ?
Sorry if it's not the place for this kind of question. Delete my message if necessary