@thexai 2022-07-10
0 ms is still the "best" but 50 ms works better than before (for 90% of people that stuck at defaults)